iOS: Regenerate VPN configuration on widget/shortcut starts
Widget, Control Center, and Siri/Shortcuts VPN starts went through
`VPNWidgetTunnelController.start()`, which only flipped `isEnabled` /
`isOnDemandEnabled` and skipped the `setupAndSave()` step the in-app
controller runs. As a result, route and exclusion settings the user
changed in Settings (enforceRoutes, includeAllNetworks,
excludeLocalNetworks, excludeAPNs, excludeCellularServices,
excludeDeviceCommunication) never reached the system VPN profile until
the user started the VPN from inside the app.

Lift the protocol-config setup into a shared
`NETunnelProviderManager.applyDuckDuckGoConfiguration(from:)` extension
in the VPN package, and call it from both the in-app controller and
the widget controller. The in-app path is functionally unchanged; the
widget/shortcut paths now regenerate the configuration from current
`VPNSettings` on every start.
