+++ iOS/DuckDuckGo/VPNShortcutIntents.swift
@@ -88,7 +88,7 @@ struct EnableVPNIntent: ForegroundContinuableIntent {
             DailyPixel.fireDailyAndCount(pixel: .vpnShortcutConnectAttempt)
 
             let controller = VPNWidgetTunnelController()
-            try await controller.start()
+            try await controller.start(settings: VPNSettings(defaults: .networkProtectionGroupDefaults))
 
             await VPNSnoozeLiveActivityManager().endSnoozeActivity()
             VPNReloadStatusWidgets()
